made corrections to task 2

+++ 0x07-pointers_arrays_strings/2-strchr.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stddef.h>
 #include "main.h"
 
 /**
@@ -11,7 +12,7 @@
 
 char *_strchr(char *s, char c)
 {
-	int a;
+	unsigned short int a;
 
 	for (a = 0; s[a] != '\0' ; a++)
 	{
